requirements/base.txt
python-consul~=0.6.0
# 'requests' and 'six' sould be installed as 'python-consul' dependencies; we just make sure
requests
six
